Extract SiteBuilder, SearchTest from PagesApiTest
