+++ _test/pages_api_test.rb
@@ -15,21 +15,13 @@
 # @author Aidan Feldman (aidan.feldman@gsa.gov)
 
 require_relative 'test_helper'
+require_relative 'site_builder'
 
 require 'minitest/autorun'
 
 module Hub
   class PagesApiTest < ::Minitest::Test
-    # before all
-    # TODO isolate the tests better
-    BUILD_DIR = File.join(Dir.pwd, '_test', 'tmp')
-    unless system(
-      "bundle exec jekyll build --destination #{BUILD_DIR} --trace",
-      {:out => '/dev/null', :err =>STDERR})
-      STDERR.puts "\n***\nSite failed to build for pages_api_test\n***\n"
-      exit $?.exitstatus
-    end
-    PATH = File.join(BUILD_DIR, 'api', 'v1', 'pages.json')
+    PATH = File.join(SiteBuilder::BUILD_DIR, 'api', 'v1', 'pages.json')
 
     def read_json(path)
       contents = File.read(path)
@@ -75,11 +67,6 @@ module Hub
 
     def test_no_posts_are_untitled
       assert_empty posts_with_title_called_untitled.map { |hash| hash['url'] }
-    end
-
-    def test_index_built
-      assert(File.exist?(File.join(BUILD_DIR, 'index.json')),
-        "Serialized lunr.js index doesn't exist")
     end
   end
 end
